requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
redis>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
celery>=5.3.0
flower>=2.0.0
psutil>=5.9.0
nltk>=3.8.0
lxml>=4.9.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
schedule>=1.2.0
APScheduler>=3.10.0
watchdog>=3.0.0
colorlog>=6.7.0
python-multipart>=0.0.6